requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
astroid==3.3.11
attrs==25.3.0
blinker==1.9.0
causal-learn==0.1.4.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
choreographer==1.1.1
clarabel==0.11.1
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.3
coverage==7.10.5
cvxpy==1.4.4
cycler==0.12.1
Cython==3.1.4
dash==3.2.0
dash-bootstrap-components==2.0.4
dill==0.4.0
directory_tree==1.0.0
dowhy==0.13
duckdb==1.3.2
econml==0.16.0
ecos==2.0.14
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
graphviz==0.21
hdbscan==0.8.40
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kaleido==1.1.0
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.44.0
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mccabe==0.7.0
momentchi2==0.1.8
mpmath==1.3.0
multidict==6.6.4
narwhals==2.2.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
orjson==3.11.3
osqp==1.0.4
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
propcache==0.3.2
psutil==7.0.0
pybind11==3.0.1
pycparser==2.23
pycryptodome==3.23.0
pydot==4.0.1
Pygments==2.19.2
pylint==3.3.8
pynndescent==0.5.13
pyparsing==3.2.3
pytest==8.4.1
pytest-cov==6.2.1
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
retrying==1.4.2
scikit-learn==1.6.1
scipy==1.15.3
scs==3.2.8
seaborn==0.13.2
shap==0.48.0
simplejson==3.20.1
six==1.17.0
slicer==0.0.8
sparse==0.17.0
SQLAlchemy==2.0.43
statsmodels==0.14.5
sympy==1.14.0
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.8.0
torch-geometric==2.6.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
Werkzeug==3.1.3
xgboost==3.0.4
yarl==1.20.1
zipp==3.23.0
